refactor: Move server code directory.
